Copy amttool from amtterm project
This commit imports the amttool script (renamed to 'amttool.pl' for
clarity) from the amtterm project: git://git.kraxel.org/amtterm

The "amtterm" package should be installed on the system to pull in what
system dependencies this script has. However, the stock script is not
easy to use; the following commit will improve it slightly, which is why
it is imported here rather than used from the standard system path.

Change-Id: I6cfb4c804af00cfeb364017566e21b9c34cb6316